Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 9ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 241115/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.cpp
@@ -30,7 +30,7 @@ int main() {
     // 여기에 코드를 작성해주세요.
     cin >> n; long long a, b; cin >> a >> b;
     int mi = n; int ma = 0;
-    for(int i=a;i<=b;i++){
+    for(long long i=a;i<=b;i++){
         int res = search(i);
         mi = min(mi,res);
         ma = max(ma,res);
